+++ src/mercury.h
@@ -61,7 +61,7 @@ struct mercury_config {
     int adaptive;                   /* adaptively accept/skip packets for PCAP output */
     bool output_block;              /* use blocking output                            */
     size_t stats_rotation_duration; /* number of seconds between stats file rotation  */
-    size_t out_rotation_duration;   /* number of seconds between json file rotation  */}
+    uint64_t out_rotation_duration; /* number of seconds between json file rotation  */}
 ;
 
 #define mercury_config_init() { NULL, NULL, NULL, NULL, NULL, NULL, O_EXCL, (char *)"w", 0, 8, 1, 0, NULL, 1, 0, 0, 0, false, 300, 0 }
